+++ wisent/core/activations/extraction_strategy.py
@@ -402,7 +402,7 @@ def extract_activation(
         # Position-weighted mean (earlier tokens weighted more)
         if num_answer_tokens > 0 and seq_len > num_answer_tokens:
             answer_hidden = hidden_states[-num_answer_tokens-1:-1]
-            weights = torch.exp(-torch.arange(answer_hidden.shape[0], dtype=torch.float32, device=answer_hidden.device) * 0.5)
+            weights = torch.exp(-torch.arange(answer_hidden.shape[0], dtype=answer_hidden.dtype, device=answer_hidden.device) * 0.5)
             weights = weights / weights.sum()
             return (answer_hidden * weights.unsqueeze(1)).sum(dim=0)
         return hidden_states[-1]
